refactor(editor): moving files such that we adhere to a "folder-by-feature" structure.
