+++ Assets/Code/NonebNi/EditorComponent/Entities/Skills.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ced7884e4326465198d1a24e4fb322ea
+timeCreated: 1685385623
