+++ Assets/Code/NonebNi/EditorComponent/Entities/Tile/TileDataSource.cs
@@ -1,7 +1,7 @@
 ﻿using NonebNi.Core.Tiles;
 using UnityEngine;
 
-namespace NonebNi.LevelEditor.Level.Data
+namespace NonebNi.EditorComponent.Entities.Tile
 {
     [CreateAssetMenu(menuName = "Data/TileData", fileName = "TileData")]
     public class TileDataSource : ScriptableObject
