--- Assets/Code/NonebNi/LevelEditor/Level/Data.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3f322cba512e4597aa31c433033ca1e9
-timeCreated: 1639915204
