+++ Assets/Code/NonebNi/LevelEditor/Level/Maps/EditorNode.Serialization.cs
