+++ Assets/Code/NonebNi/LevelEditor/Level/Maps/EditorNode.cs
